@@ -7,42 +7,47 @@ issue_tracker: https://github.com/nlfiedler/tanuki/issues
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  equatable: ^1.2.5
-  file_picker: ^2.0.11
-  flutter_bloc: ^6.1.1
+  equatable: ^2.0.0
+  file_picker: ^3.0.1
+  flutter_bloc: ^7.0.0
   flutter_chips_input: ^1.9.5
   # use file_chooser vs file_picker on desktop to avoid installing golang
   # (temporary solution for desktop file chooser, no package yet available)
+  # TODO: replace this with https://github.com/google/flutter-desktop-embedding/tree/master/plugins/file_selector
+  # TODO: published as https://pub.dev/packages/file_selector
   file_chooser:
     git:
       url: git://github.com/google/flutter-desktop-embedding.git
       path: plugins/file_chooser
       ref: 927f8cbc09b35d85245c095f2db8df9b186f6618
-  flutter_dropzone: ^1.0.9
+  flutter_dropzone: ^2.0.1
   flutter_form_builder: ^6.0.0
-  flutter_riverpod: ^0.14.0+2
-  graphql: ^3.1.0
-  http: ^0.12.2
+  flutter_riverpod: ^0.14.0+3
+  graphql: ^5.0.0-nullsafety.4
+  http: ^0.13.2
   intl: ^0.17.0
-  mime: ^0.9.7
-  oxidized: ^3.1.0
+  mime: ^1.0.0
+  oxidized: ^4.2.0
   path: ^1.7.0
-  url_launcher: ^5.7.8
-  video_player: ^0.11.1+2
+  url_launcher: ^6.0.3
+  video_player: ^2.1.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  bloc_test: ^7.1.0
-  environment_config: ^2.2.3
-  mockito: ^4.1.4
+  bloc_test: ^8.0.0
+  build_runner: ^2.0.1
+  # build_test depends on an older test_core, for now this is not available
+  # build_test: ^2.1.0
+  environment_config: ^3.0.0
+  mockito: ^5.0.7
   pedantic: ^1.8.0
 
 flutter:
