@@ -1,18 +1,17 @@
 //
 // Copyright (c) 2020 Nathan Fiedler
 //
-import 'package:meta/meta.dart';
 import 'package:oxidized/oxidized.dart';
 import 'package:tanuki/core/domain/entities/input.dart';
 
 class AssetInputModel extends AssetInput {
   AssetInputModel({
-    @required List<String> tags,
-    @required Option<String> caption,
-    @required Option<String> location,
-    @required Option<DateTime> datetime,
-    @required Option<String> mimetype,
-    @required Option<String> filename,
+    required List<String> tags,
+    required Option<String> caption,
+    required Option<String> location,
+    required Option<DateTime> datetime,
+    required Option<String> mimetype,
+    required Option<String> filename,
   }) : super(
           tags: tags,
           caption: caption,
@@ -35,12 +34,12 @@ class AssetInputModel extends AssetInput {
 
   factory AssetInputModel.fromJson(Map<String, dynamic> json) {
     final List<String> tags = List.from(json['tags'].map((t) => t.toString()));
-    final Option<String> caption = Option.some(json['caption']);
-    final Option<String> location = Option.some(json['location']);
-    final Option<String> filename = Option.some(json['filename']);
-    final Option<String> mimetype = Option.some(json['mimetype']);
-    final datetime = Option.some(json['datetime']).map(
-      (v) => DateTime.parse(v),
+    final Option<String> caption = Option.from(json['caption']);
+    final Option<String> location = Option.from(json['location']);
+    final Option<String> filename = Option.from(json['filename']);
+    final Option<String> mimetype = Option.from(json['mimetype']);
+    final datetime = Option.from(json['datetime']).map(
+      (v) => DateTime.parse(v as String),
     );
     return AssetInputModel(
       tags: tags,
@@ -55,19 +54,19 @@ class AssetInputModel extends AssetInput {
   Map<String, dynamic> toJson() {
     return {
       'tags': tags,
-      'caption': caption.unwrapOr(null),
-      'location': location.unwrapOr(null),
+      'caption': caption.unwrapOr(''),
+      'location': location.unwrapOr(''),
       'datetime': datetime.mapOr((v) => v.toIso8601String(), null),
-      'mimetype': mimetype.unwrapOr(null),
-      'filename': filename.unwrapOr(null),
+      'mimetype': mimetype.unwrapOr(''),
+      'filename': filename.unwrapOr(''),
     };
   }
 }
 
 class AssetInputIdModel extends AssetInputId {
   AssetInputIdModel({
-    @required String id,
-    @required AssetInput input,
+    required String id,
+    required AssetInput input,
   }) : super(id: id, input: input);
 
   factory AssetInputIdModel.from(AssetInputId inputId) {
