@@ -1,7 +1,6 @@
 //
 // Copyright (c) 2020 Nathan Fiedler
 //
-import 'package:meta/meta.dart';
 import 'package:oxidized/oxidized.dart';
 import 'package:tanuki/core/data/sources/entity_remote_data_source.dart';
 import 'package:tanuki/core/domain/entities/asset.dart';
@@ -16,16 +15,13 @@ class EntityRepositoryImpl extends EntityRepository {
   final EntityRemoteDataSource remoteDataSource;
 
   EntityRepositoryImpl({
-    @required this.remoteDataSource,
+    required this.remoteDataSource,
   });
 
   @override
   Future<Result<int, Failure>> bulkUpdate(List<AssetInputId> assets) async {
     try {
       final results = await remoteDataSource.bulkUpdate(assets);
-      if (results == null) {
-        return Err(ServerFailure('got null result for query'));
-      }
       return Ok(results);
     } on ServerException catch (e) {
       return Err(ServerFailure(e.toString()));
@@ -36,9 +32,6 @@ class EntityRepositoryImpl extends EntityRepository {
   Future<Result<List<Location>, Failure>> getAllLocations() async {
     try {
       final locations = await remoteDataSource.getAllLocations();
-      if (locations == null) {
-        return Err(ServerFailure('got null result for locations'));
-      }
       return Ok(locations);
     } on ServerException catch (e) {
       return Err(ServerFailure(e.toString()));
@@ -49,9 +42,6 @@ class EntityRepositoryImpl extends EntityRepository {
   Future<Result<List<Tag>, Failure>> getAllTags() async {
     try {
       final tags = await remoteDataSource.getAllTags();
-      if (tags == null) {
-        return Err(ServerFailure('got null result for tags'));
-      }
       return Ok(tags);
     } on ServerException catch (e) {
       return Err(ServerFailure(e.toString()));
@@ -62,9 +52,6 @@ class EntityRepositoryImpl extends EntityRepository {
   Future<Result<List<Year>, Failure>> getAllYears() async {
     try {
       final years = await remoteDataSource.getAllYears();
-      if (years == null) {
-        return Err(ServerFailure('got null result for years'));
-      }
       return Ok(years);
     } on ServerException catch (e) {
       return Err(ServerFailure(e.toString()));
@@ -88,9 +75,6 @@ class EntityRepositoryImpl extends EntityRepository {
   Future<Result<int, Failure>> getAssetCount() async {
     try {
       final count = await remoteDataSource.getAssetCount();
-      if (count == null) {
-        return Err(ServerFailure('got null result for count'));
-      }
       return Ok(count);
     } on ServerException catch (e) {
       return Err(ServerFailure(e.toString()));
