@@ -2,7 +2,6 @@
 // Copyright (c) 2020 Nathan Fiedler
 //
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 import 'package:oxidized/oxidized.dart';
 import 'package:tanuki/core/domain/entities/input.dart';
 import 'package:tanuki/core/domain/repositories/entity_repository.dart';
@@ -24,7 +23,7 @@ class Params extends Equatable {
   final List<AssetInputId> assets;
 
   Params({
-    @required this.assets,
+    required this.assets,
   });
 
   @override
