@@ -2,7 +2,6 @@
 // Copyright (c) 2020 Nathan Fiedler
 //
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 import 'package:oxidized/oxidized.dart';
 import 'package:tanuki/core/domain/entities/asset.dart';
 import 'package:tanuki/core/domain/repositories/entity_repository.dart';
@@ -26,7 +25,7 @@ class Params extends Equatable {
   final String assetId;
 
   Params({
-    @required this.assetId,
+    required this.assetId,
   });
 
   @override
