@@ -4,7 +4,6 @@
 import 'dart:async';
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
-import 'package:meta/meta.dart';
 import 'package:tanuki/core/domain/entities/attributes.dart';
 import 'package:tanuki/core/domain/usecases/get_all_tags.dart';
 import 'package:tanuki/core/domain/usecases/usecase.dart';
@@ -36,7 +35,7 @@ class Loading extends AllTagsState {}
 class Loaded extends AllTagsState {
   final List<Tag> tags;
 
-  Loaded({@required this.tags});
+  Loaded({required this.tags});
 
   @override
   List<Object> get props => [tags];
@@ -45,7 +44,7 @@ class Loaded extends AllTagsState {
 class Error extends AllTagsState {
   final String message;
 
-  Error({@required this.message});
+  Error({required this.message});
 
   @override
   List<Object> get props => [message];
@@ -58,7 +57,7 @@ class Error extends AllTagsState {
 class AllTagsBloc extends Bloc<AllTagsEvent, AllTagsState> {
   final GetAllTags usecase;
 
-  AllTagsBloc({this.usecase}) : super(Empty());
+  AllTagsBloc({required this.usecase}) : super(Empty());
 
   @override
   Stream<AllTagsState> mapEventToState(
