@@ -4,19 +4,20 @@
 import 'dart:typed_data';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
+import 'package:mockito/annotations.dart';
 import 'package:tanuki/core/error/exceptions.dart';
 import 'package:tanuki/core/error/failures.dart';
 import 'package:tanuki/core/data/sources/asset_remote_data_source.dart';
 import 'package:tanuki/core/data/repositories/asset_repository_impl.dart';
+import './asset_repository_impl_test.mocks.dart';
 
-class MockRemoteDataSource extends Mock implements AssetRemoteDataSource {}
-
+@GenerateMocks([AssetRemoteDataSource])
 void main() {
-  AssetRepositoryImpl repository;
-  MockRemoteDataSource mockRemoteDataSource;
+  late AssetRepositoryImpl repository;
+  late MockAssetRemoteDataSource mockRemoteDataSource;
 
   setUp(() {
-    mockRemoteDataSource = MockRemoteDataSource();
+    mockRemoteDataSource = MockAssetRemoteDataSource();
     repository = AssetRepositoryImpl(
       remoteDataSource: mockRemoteDataSource,
     );
@@ -34,19 +35,6 @@ void main() {
         verify(mockRemoteDataSource.ingestAssets());
         expect(result.unwrap(), isA<int>());
         expect(result.unwrap(), equals(123));
-      },
-    );
-
-    test(
-      'should return failure when remote data source returns null',
-      () async {
-        // arrange
-        when(mockRemoteDataSource.ingestAssets()).thenAnswer((_) async => null);
-        // act
-        final result = await repository.ingestAssets();
-        // assert
-        verify(mockRemoteDataSource.ingestAssets());
-        expect(result.err().unwrap(), isA<ServerFailure>());
       },
     );
 
@@ -81,20 +69,6 @@ void main() {
     );
 
     test(
-      'should return failure when remote data source returns null',
-      () async {
-        // arrange
-        when(mockRemoteDataSource.uploadAsset(any))
-            .thenAnswer((_) async => null);
-        // act
-        final result = await repository.uploadAsset('happy_cow.jpg');
-        // assert
-        verify(mockRemoteDataSource.uploadAsset('happy_cow.jpg'));
-        expect(result.err().unwrap(), isA<ServerFailure>());
-      },
-    );
-
-    test(
       'should return failure when remote data source is unsuccessful',
       () async {
         // arrange
@@ -123,21 +97,6 @@ void main() {
         verify(mockRemoteDataSource.uploadAssetBytes('happy.jpg', bytes));
         expect(result.unwrap(), isA<String>());
         expect(result.unwrap(), equals('asset123'));
-      },
-    );
-
-    test(
-      'should return failure when remote data source returns null',
-      () async {
-        // arrange
-        when(mockRemoteDataSource.uploadAssetBytes(any, any))
-            .thenAnswer((_) async => null);
-        // act
-        final bytes = Uint8List(0);
-        final result = await repository.uploadAssetBytes('happy.jpg', bytes);
-        // assert
-        verify(mockRemoteDataSource.uploadAssetBytes('happy.jpg', bytes));
-        expect(result.err().unwrap(), isA<ServerFailure>());
       },
     );
 
