@@ -3,17 +3,18 @@
 //
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
+import 'package:mockito/annotations.dart';
 import 'package:oxidized/oxidized.dart';
 import 'package:tanuki/core/domain/entities/attributes.dart';
 import 'package:tanuki/core/domain/repositories/entity_repository.dart';
 import 'package:tanuki/core/domain/usecases/get_all_years.dart';
 import 'package:tanuki/core/domain/usecases/usecase.dart';
+import './get_all_years_test.mocks.dart';
 
-class MockEntityRepository extends Mock implements EntityRepository {}
-
+@GenerateMocks([EntityRepository])
 void main() {
-  GetAllYears usecase;
-  MockEntityRepository mockEntityRepository;
+  late GetAllYears usecase;
+  late MockEntityRepository mockEntityRepository;
 
   final years = [
     Year(label: '2019', count: 806),
