@@ -4,17 +4,18 @@
 import 'package:bloc_test/bloc_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
+import 'package:mockito/annotations.dart';
 import 'package:oxidized/oxidized.dart';
 import 'package:tanuki/core/domain/repositories/asset_repository.dart';
 import 'package:tanuki/core/domain/usecases/ingest_assets.dart';
 import 'package:tanuki/core/error/failures.dart';
 import 'package:tanuki/features/import/preso/bloc/ingest_assets_bloc.dart';
+import './ingest_assets_bloc_test.mocks.dart';
 
-class MockAssetRepository extends Mock implements AssetRepository {}
-
+@GenerateMocks([AssetRepository])
 void main() {
-  MockAssetRepository mockAssetRepository;
-  IngestAssets usecase;
+  late MockAssetRepository mockAssetRepository;
+  late IngestAssets usecase;
 
   group('normal cases', () {
     setUp(() {
@@ -26,14 +27,14 @@ void main() {
     blocTest(
       'emits [] when nothing is added',
       build: () => IngestAssetsBloc(usecase: usecase),
-      expect: [],
+      expect: () => [],
     );
 
     blocTest(
       'emits [Processing, Finished] when ProcessUploads is added',
       build: () => IngestAssetsBloc(usecase: usecase),
-      act: (bloc) => bloc.add(ProcessUploads()),
-      expect: [
+      act: (IngestAssetsBloc bloc) => bloc.add(ProcessUploads()),
+      expect: () => [
         Processing(),
         Finished(count: 101),
       ],
@@ -51,8 +52,8 @@ void main() {
     blocTest(
       'emits [Uploading, Error] when repository returns an error',
       build: () => IngestAssetsBloc(usecase: usecase),
-      act: (bloc) => bloc.add(ProcessUploads()),
-      expect: [
+      act: (IngestAssetsBloc bloc) => bloc.add(ProcessUploads()),
+      expect: () => [
         Processing(),
         Error(message: 'ServerFailure(oh no!)'),
       ],
